requirements.txt
torch==2.3.1
transformers==4.42.4
peft==0.11.1
trl==0.9.6
bitsandbytes==0.43.1
accelerate==0.32.1
datasets==2.20.0
google-generativeai==0.8.6
python-dotenv==1.0.1
tqdm==4.66.4
